move edgeFilteringWarning tests to separate file'
